ml_backend/requirements.txt
# FastAPI and Server
fastapi
uvicorn[standard]
python-dotenv
asyncpg

# ML Models
prophet
PyPortfolioOpt
scikit-learn
pandas
numpy

# Data Fetching
httpx

# Model Management
joblib 